+++ features/step_definitions/additional_cli_steps.rb
@@ -18,3 +18,13 @@ end
 Then /^the file "([^"]*)" should contain:$/ do |file, partial_content|
   check_file_content(file, partial_content, true)
 end
+
+Then /^the backtrace\-normalized output should contain:$/ do |partial_output|
+  # ruby 1.9 includes additional stuff in the backtrace,
+  # so we need to normalize it to compare it with our expected output.
+  normalized_output = combined_output.split("\n").map do |line|
+    line =~ /(^\s+# [^:]+:\d+)/ ? $1 : line # http://rubular.com/r/zDD7DdWyzF
+  end.join("\n")
+
+  normalized_output.should =~ compile_and_escape(partial_output)
+end